move auth stuff to libs
