+++ app/main.tsx
@@ -1,3 +1,4 @@
+import { AuthProvider } from '@camp/auth';
 import { ApolloProvider } from '@camp/data-layer';
 import { ThemeProvider } from '@camp/design';
 import { ModalsProvider } from '@mantine/modals';
@@ -5,7 +6,6 @@ import { NotificationsProvider } from '@mantine/notifications';
 import { StrictMode } from 'react';
 import ReactDOM from 'react-dom/client';
 
-import { AuthProvider } from './AuthProvider';
 import { Routes } from './Routes';
 
 const root = ReactDOM.createRoot(document.getElementById('root')!);
