+++ libs/auth/README.md
@@ -1,0 +1,9 @@
+# Auth
+
+## Concerns
+
+* A provider using auth0 for authorization
+
+## Dependency Rules
+
+* The application should't import `@auth0/auth0-react` directly.
