--- libs/design/components/Guards/index.ts
@@ -1,2 +1,0 @@
-export * from './AuthGuard';
-export * from './GuestGuard';
